% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acceptence_Probability.R
\name{Acceptence_Probability}
\alias{Acceptence_Probability}
\title{Acceptence Probability}
\usage{
Acceptence_Probability(
  x,
  Tess,
  Dim,
  j,
  R_ijOld,
  n_ijOld,
  R_ijNew,
  n_ijNew,
  SigmaSquared,
  Modification,
  SigmaSquaredMu,
  Omega,
  lambda_rate
)
}
\arguments{
\item{x}{The covariate matrix.}

\item{Tess}{The tessellation.}

\item{Dim}{The dimensions of the tessellation.}

\item{j}{The index of the tessellation.}

\item{R_ijOld}{The residuals of the old tessellation.}

\item{n_ijOld}{The number of observations in each cell of the old tessellation.}

\item{R_ijNew}{The residuals of the new tessellation.}

\item{n_ijNew}{The number of observations in each cell of the new tessellation.}

\item{SigmaSquared}{The variance of the residuals.}

\item{Modification}{The type of modification.}

\item{SigmaSquaredMu}{The variance of the random effects.}

\item{Omega}{The prior probability of adding a dimension.}

\item{lambda_rate}{The rate of the Poisson distribution for the number of centers.}
}
\value{
The acceptence probability.
}
\description{
Calculates the acceptence rate of the proposed tessellation.
}
