% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortes-package.R
\docType{package}
\name{AddiVortes-package}
\alias{AddiVortes-package}
\title{AddiVortes: Bayesian Additive Voronoi Tessellations for Machine Learning}
\description{
AddiVortes implements Bayesian Additive Voronoi Tessellation models for
machine learning regression, classification and non-parametric statistical
modelling. This package provides a flexible alternative to BART (Bayesian
Additive Regression Trees), using Voronoi tessellations instead of trees for
spatial partitioning. The method is particularly effective for spatial data
analysis, complex function approximation, and Bayesian regression and
classification.
}
\details{
Key features include:
\itemize{
\item Machine learning regression and classification with Bayesian inference
\item Binary and multinomial probit models with latent-variable data
augmentation
\item Alternative to BART using Voronoi tessellations
\item Spatial data analysis and modelling
\item Non-parametric regression capabilities
\item Complex function approximation
\item Uncertainty quantification through posterior inference
}
}
\references{
Stone, A. and Gosling, J.P. (2025). AddiVortes: (Bayesian) additive Voronoi
tessellations. Journal of Computational and Graphical Statistics.

Stone, A.J., Ogundimu, E. and Gosling, J.P. (2026). Binary AddiVortes:
(Bayesian) Additive Voronoi Tessellations for Binary Classification with an
application to Predicting Home Mortgage Application Outcomes.

Albert, J.H. and Chib, S. (1993). Bayesian analysis of binary and
polychotomous response data. Journal of the American Statistical Association.

Kindo, B.P., Wang, H. and Peña, E.A. (2016). Multinomial probit Bayesian
additive regression trees. Stat.
}
\seealso{
\url{https://johnpaulgosling.github.io/AddiVortes/}
}
\author{
\strong{Maintainer}: John Paul Gosling \email{john-paul.gosling@durham.ac.uk} (\href{https://orcid.org/0000-0002-4072-3022}{ORCID})

Authors:
\itemize{
  \item John Paul Gosling \email{john-paul.gosling@durham.ac.uk} (\href{https://orcid.org/0000-0002-4072-3022}{ORCID})
  \item Adam Stone \email{adam.stone2@durham.ac.uk} (\href{https://orcid.org/0009-0004-0058-6117}{ORCID})
  \item Andrew Iskauskas \email{andrew.iskauskas@durham.ac.uk} (\href{https://orcid.org/0000-0003-2825-3651}{ORCID})
  \item Leo Thomson \email{leo@feasibly.co.uk}
}

}
\keyword{BART}
\keyword{bayesian}
\keyword{classification}
\keyword{machine-learning}
\keyword{package}
\keyword{regression}
\keyword{spatial}
\keyword{tessellation}
