% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Boston}
\alias{Boston}
\title{Boston Dataset}
\format{
\subsection{\code{Boston}}{

A data.frame with 506 rows and 14 columns:
\describe{
\item{x1-x13}{The covariates in the data}
\item{y}{The response variable}
}
}
}
\source{
Harrison, D. and Rubinfeld, D.L. `Hedonic prices and the demand for
clean air', J. Environ. Economics & Management, vol. 5, 81-102, 1978
}
\usage{
Boston
}
\description{
The Boston Housing Dataset, derived from information from the U.S. Census
Service.
}
\keyword{datasets}
