% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Residuals.R
\name{Calculate_Residuals}
\alias{Calculate_Residuals}
\title{Calculate_Residuals}
\usage{
Calculate_Residuals(
  y,
  x,
  j,
  SumOfAllTess,
  Tess,
  Dim,
  Pred,
  TessStar,
  DimStar,
  LastTessPred
)
}
\arguments{
\item{y}{The outcome variable.}

\item{x}{The covariate matrix.}

\item{j}{The index of the tessellation.}

\item{SumOfAllTess}{The sum of all tessellations.}

\item{Tess}{The tessellation.}

\item{Dim}{The dimensions of the tessellation.}

\item{Pred}{The prediction set.}

\item{TessStar}{The proposed tessellation.}

\item{DimStar}{The proposed dimensions.}

\item{LastTessPred}{The last tessellation prediction.}
}
\value{
A list containing the residuals of the old tessellation, the number of observations in each cell of the old tessellation, the residuals of the new tessellation, the number of observations in each cell of the new tessellation, the sum of all tessellations, the indexes of the proposed tessellation, and the indexes of the original tessellation.
}
\description{
A function that calculates the n-vector of partial residuals derived from a
fitting process that excludes the jth tessellation.
}
