% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Indexes.R
\name{Cell_Indexes}
\alias{Cell_Indexes}
\title{Cell_Indexes}
\usage{
Cell_Indexes(x, Tess, Dim)
}
\arguments{
\item{x}{The covariate matrix.}

\item{Tess}{The tessellation.}

\item{Dim}{The dimensions of the tessellation.}
}
\value{
The row (the center) of the tessellation that each observation falls within.
}
\description{
A function that gives the row (the center) of the tessellation that each observation falls within.
}
