% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cylinder}
\alias{Cylinder}
\title{Cylindrical Dataset}
\format{
\subsection{\code{Cylinder}}{

A data.frame with 400 rows and 3 columns:
\describe{
\item{z}{Height}
\item{theta}{Polar angle}
\item{Y}{Response}
}
}
}
\usage{
Cylinder
}
\description{
Synthetic data generated for cylindrical parameters. The response is given by
the function f(z, theta) = sin(z)cos(theta) + z*sin(theta)^2, with added noise.
}
\keyword{datasets}
