% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Fitting_Function.R
\name{Fitting_Function}
\alias{Fitting_Function}
\title{Fitting function}
\usage{
Fitting_Function(lambda, q, nu, sigmaSquared_hat)
}
\arguments{
\item{lambda}{The lambda value.}

\item{q}{The q value.}

\item{nu}{The nu value.}

\item{sigmaSquared_hat}{The sigma squared hat value.}
}
\value{
The squared difference between sigma squared hat and the inverse gamma function.
}
\description{
A function that calculates the squared difference between sigma squared hat
and the inverse gamma function.
}
