% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_FittingFunction.R
\name{fittingFunction}
\alias{fittingFunction}
\title{Fitting function}
\usage{
fittingFunction(lambda, q, nu, sigmaSquaredHat)
}
\arguments{
\item{lambda}{The lambda value.}

\item{q}{The q value.}

\item{nu}{The nu value.}

\item{sigmaSquaredHat}{The sigma squared hat value.}
}
\value{
The squared difference between sigma squared hat and the inverse gamma function.
}
\description{
A function that calculates the squared difference between sigma squared hat
and the inverse gamma function.
}
