% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortesFit.R
\name{new_AddiVortesFit}
\alias{new_AddiVortesFit}
\title{Create an AddiVortesFit Object}
\usage{
new_AddiVortesFit(
  posteriorTess,
  posteriorDim,
  posteriorPred,
  xCentres,
  xRanges,
  yCentre,
  yRange,
  inSampleRmse
)
}
\arguments{
\item{posteriorTess}{A list of the posterior samples of the tessellations.}

\item{posteriorDim}{A list of the posterior samples of the dimensions.}

\item{posteriorPred}{A list of the posterior samples of the predictions.}

\item{xCentres}{The centres of the covariates.}

\item{xRanges}{The ranges of the covariates.}

\item{yCentre}{The centre of the output values.}

\item{yRange}{The range of the output values.}

\item{inSampleRmse}{The in-sample RMSE.}
}
\value{
An object of class AddiVortesFit.
}
\description{
A constructor for the AddiVortesFit class.
}
