% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictAddiVortes.R
\name{PredictAddiVortes}
\alias{PredictAddiVortes}
\title{PredictAddiVortes}
\usage{
PredictAddiVortes(addivortes_model_fit, x_new, y_new)
}
\arguments{
\item{addivortes_model_fit}{The 'AddiVortes_model' list object returned by the
AddiVortes function. This contains posterior samples
and scaling parameters.}

\item{x_new}{A matrix of the covariates for the new test set. The number of
columns must match the original training data.}

\item{y_new}{A vector of the true output values for the new test set. Must have
the same number of observations as rows in \code{x_new}.}
}
\value{
The RMSE value for the new test samples. Returns NA if the model
contains no posterior samples.
}
\description{
Utilises a fitted AddiVortes model to make predictions on new test data and
calculates the Root Mean Squared Error (RMSE). This function aims to
reproduce the Out-of-Sample RMSE calculation performed internally by the
AddiVortes fitting function.
}
\details{
This function requires the helper functions \code{apply_scaling_internal} and
\code{TestSet_Prediction} (used by the original \code{AddiVortes} function) to be
available in the R environment.
}
