% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print.thiessen_option}
\alias{print.thiessen_option}
\title{Print a component option}
\usage{
\method{print}{thiessen_option}(x, ...)
}
\arguments{
\item{x}{An option from \code{\link[=soft_membership]{soft_membership()}}, \code{\link[=weighted_inclusion]{weighted_inclusion()}} or
\code{\link[=dart_inclusion]{dart_inclusion()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a component option
}
\examples{
\dontshow{if (core_experimental()) withAutoprint(\{ # examplesIf}
print(dart_inclusion())
\dontshow{\}) # examplesIf}
}
