% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{print.thiessen_sampler}
\alias{print.thiessen_sampler}
\title{Print a sampler}
\usage{
\method{print}{thiessen_sampler}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"thiessen_sampler"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a sampler
}
\examples{
design <- matrix(seq(0, 1, length.out = 40), ncol = 1)
print(thiessen_sampler(design, design[, 1]^2,
                       thiessen_control(tessellations = 10), seed = 1))
}
