% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Propose_Tessellation.R
\name{Propose_Tessellation}
\alias{Propose_Tessellation}
\title{Propose_Tessellation}
\usage{
Propose_Tessellation(x, j, Tess, Dim, var)
}
\arguments{
\item{x}{A matrix of covariates.}

\item{j}{The index of the tessellation to be modified.}

\item{Tess}{A list of matrices of the tessellations.}

\item{Dim}{A list of vectors of the dimensions of the tessellations.}

\item{var}{The variance of the normal distribution used to sample new coordinates.}
}
\value{
A list containing the new tessellation, the new dimension matrix and the modification made.
}
\description{
Propose a new tessellation by adding, removing, changing or swopping dimensions or centers.
}
