% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstantools.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{posterior_predict}
\alias{posterior_epred}
\alias{log_lik}
\alias{predictive_interval}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rstantools}{\code{\link[rstantools:log_lik]{log_lik()}}, \code{\link[rstantools:posterior_epred]{posterior_epred()}}, \code{\link[rstantools:posterior_predict]{posterior_predict()}}, \code{\link[rstantools:predictive_interval]{predictive_interval()}}}
}}

