% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_mu_values.R
\name{Sample_mu_values}
\alias{Sample_mu_values}
\title{Sample_mu_values}
\usage{
Sample_mu_values(j, Tess, R_ijNew, n_ijNew, sigmaSquaredMu, SigmaSquared)
}
\arguments{
\item{j}{The index of the tessellation.}

\item{Tess}{The tessellation.}

\item{R_ijNew}{The new R_ij values.}

\item{n_ijNew}{The new n_ij values.}

\item{sigmaSquaredMu}{The sigma squared mu value.}

\item{SigmaSquared}{The Sigma squared value.}
}
\value{
The new output values for the new tessellation.
}
\description{
This function samples the new mean output values for
the proposed tessellation.
}
