% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Sigma_Squared.R
\name{Sample_Sigma_Squared}
\alias{Sample_Sigma_Squared}
\title{Sample Sigma Squared}
\usage{
Sample_Sigma_Squared(yScaled, nu, lambda, SumOfAllTess)
}
\arguments{
\item{yScaled}{A vector of scaled data.}

\item{nu}{The prior degrees of freedom.}

\item{lambda}{The prior scale parameter.}

\item{SumOfAllTess}{The sum of all the tessellations.}
}
\value{
SigmaSquared The sampled sigma squared.
}
\description{
This function samples sigma squared from the inverse gamma distribution.
}
