% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_SampleMuValues.R
\name{sampleMuValues}
\alias{sampleMuValues}
\title{sampleMuVvalues}
\usage{
sampleMuValues(j, Tess, R_ijNew, n_ijNew, sigmaSquaredMu, sigmaSquared)
}
\arguments{
\item{j}{The index of the tessellation.}

\item{Tess}{The tessellation.}

\item{R_ijNew}{The new R_ij values.}

\item{n_ijNew}{The new n_ij values.}

\item{sigmaSquaredMu}{The sigma squared mu value.}

\item{sigmaSquared}{The Sigma squared value.}
}
\value{
The new output values for the new tessellation.
}
\description{
This function samples the new mean output values for
the proposed tessellation.
}
