% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_SampleSigmaSquared.R
\name{sampleSigmaSquared}
\alias{sampleSigmaSquared}
\title{Sample Sigma Squared}
\usage{
sampleSigmaSquared(yScaled, nu, lambda, SumOfAllTess)
}
\arguments{
\item{yScaled}{A vector of scaled data.}

\item{nu}{The prior degrees of freedom.}

\item{lambda}{The prior scale parameter.}

\item{SumOfAllTess}{The sum of all the tessellations.}
}
\value{
sigmaSquared The sampled sigma squared.
}
\description{
This function samples sigma squared from the inverse gamma distribution.
}
