% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortesFit.R
\name{summary.AddiVortesFit}
\alias{summary.AddiVortesFit}
\title{Summary Method for AddiVortesFit}
\usage{
\method{summary}{AddiVortesFit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{AddiVortesFit}, typically the result of a
call to \code{AddiVortes()}.}

\item{...}{Further arguments passed to or from other methods (currently
unused).}
}
\value{
The function is called for its side effect of printing detailed model
information and returns the input object \code{object} invisibly.
}
\description{
Provides a detailed summary of a fitted \code{AddiVortesFit} object, including
more comprehensive information than the print method.
}
