% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSetPrediction.R
\name{testSetPrediction}
\alias{testSetPrediction}
\title{Test Set Prediction}
\usage{
testSetPrediction(x, m, Tess, Dim, Pred)
}
\arguments{
\item{x}{A matrix with the observations.}

\item{m}{The number of iterations in AddiVortes.}

\item{Tess}{A list with the tessellations.}

\item{Dim}{A list with the dimensions of the tessellations.}

\item{Pred}{A list with the predictions.}
}
\value{
A vector with the predictions.
}
\description{
A function that derives a prediction for the output variable
give an observation, for one iteration in AddiVortes.
}
