% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortesFit.R
\name{traceplots}
\alias{traceplots}
\title{Trace plots}
\usage{
traceplots(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
S3 generic for MCMC trace plots of fitted AddiVortes models. Users should
call \code{traceplots(fit)}; see \code{\link[=traceplots.AddiVortes]{traceplots.AddiVortes()}} for the method
documentation.
}
\seealso{
\code{\link[=traceplots.AddiVortes]{traceplots.AddiVortes()}}
}
\keyword{internal}
