% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Weather}
\alias{Weather}
\title{Weather Dataset}
\format{
\subsection{\code{Weather}}{

A data.frame with 2000 rows and 3 columns:
\describe{
\item{theta}{Polar angle}
\item{phi}{Azimuthal angle}
\item{Y}{Response variable}
}
}
}
\source{
https://disc.gsfc.nasa.gov/datasets/GPM_1CGPMGMI_07/summary?keywords=10.5067\%2FGPM\%2FGMI\%2FGPM\%2F1C\%2F07
}
\usage{
Weather
}
\description{
A subset of data collected as part of the GES DISC datasets for calibrated
brightness temperatures.
}
\keyword{datasets}
