% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortes-package.R
\docType{package}
\name{AddiVortes-package}
\alias{AddiVortes-package}
\title{AddiVortes: (Bayesian) Additive Voronoi Tessellations}
\description{
Implements the Bayesian Additive Voronoi Tessellation (AddiVortes) model for non-parametric regression and classification. This package allows users to fit the model estimate posterior distributions and visualise the resulting tessellations. It is particularly useful for spatial data analysis and complex function approximation where the underlying structure is unknown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://johnpaulgosling.github.io/AddiVortes/}
  \item Report bugs at \url{https://github.com/johnpaulgosling/AddiVortes/issues}
}

}
\author{
\strong{Maintainer}: Adam Stone \email{adam.stone2@durham.ac.uk} (\href{https://orcid.org/0009-0004-0058-6117}{ORCID})

Authors:
\itemize{
  \item John Paul Gosling \email{john-paul.gosling@durham.ac.uk} (\href{https://orcid.org/0000-0002-4072-3022}{ORCID})
}

}
\keyword{internal}
