% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortes.R
\name{AddiVortes}
\alias{AddiVortes}
\title{AddiVortes}
\usage{
AddiVortes(
  y,
  x,
  m = 200,
  max_iter = 1200,
  burn_in = 200,
  nu = 6,
  q = 0.85,
  k = 3,
  sd = 0.8,
  Omega = 3,
  lambda_rate = 25,
  YTest,
  XTest,
  IntialSigma = "Linear",
  thinning = 1
)
}
\arguments{
\item{y}{A vector of the output values.}

\item{x}{A matrix of the covariates.}

\item{m}{The number of tessellations.}

\item{max_iter}{The number of iterations.}

\item{burn_in}{The number of burn in iterations.}

\item{nu}{The degrees of freedom.}

\item{q}{The quantile.}

\item{k}{The number of centers.}

\item{sd}{The standard deviation.}

\item{Omega}{The prior probability of adding a dimension.}

\item{lambda_rate}{The rate of the Poisson distribution for the number of centers.}

\item{YTest}{A vector of the output values for the test set.}

\item{XTest}{A matrix of the covariates for the test set.}

\item{IntialSigma}{The method used to calculate the initial sigma.}

\item{thinning}{The thinning rate.}
}
\value{
The RMSE value for the test samples.
}
\description{
The AddiVortes function is a Bayesian nonparametric regression model that uses a tessellation to model the relationship between the covariates and the output values. The model uses a backfitting algorithm to sample from the posterior distribution of the output values for each tessellation. The function returns the RMSE value for the test samples.
}
