% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellIndexes.R
\name{cellIndices}
\alias{cellIndices}
\title{Cell Indices}
\usage{
cellIndices(x, Tess, Dim)
}
\arguments{
\item{x}{The covariate matrix.}

\item{Tess}{The tessellation.}

\item{Dim}{The dimensions of the tessellation.}
}
\value{
The row (the center) of the tessellation that each observation falls within.
}
\description{
A function that gives the row (the center) of the tessellation
that each observation falls within.
}
