% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_function.R
\name{fitting_function}
\alias{fitting_function}
\title{Fitting function}
\usage{
fitting_function(lambda, q, nu, sigmaSquared_hat)
}
\arguments{
\item{lambda}{The lambda value.}

\item{q}{The q value.}

\item{nu}{The nu value.}

\item{sigmaSquared_hat}{The sigma squared hat value.}
}
\value{
The squared difference between sigma squared hat and the inverse gamma function.
}
\description{
A function that calculates the squared difference between sigma squared hat and the inverse gamma function.
}
