% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortesFit.R
\name{new_AddiVortes}
\alias{new_AddiVortes}
\title{Create an AddiVortes Object}
\usage{
new_AddiVortes(
  posteriorTess,
  posteriorDim,
  posteriorSigma,
  posteriorPred,
  xCentres,
  xRanges,
  yCentre,
  yRange,
  inSampleRmse,
  metric = "E",
  catEncoding = NULL
)
}
\arguments{
\item{posteriorTess}{A list of the posterior samples of the tessellations.}

\item{posteriorDim}{A list of the posterior samples of the dimensions.}

\item{posteriorSigma}{A list of the posterior samples of the error variance.}

\item{posteriorPred}{A list of the posterior samples of the predictions.}

\item{xCentres}{The centres of the covariates.}

\item{xRanges}{The ranges of the covariates.}

\item{yCentre}{The centre of the output values.}

\item{yRange}{The range of the output values.}

\item{inSampleRmse}{The in-sample RMSE.}

\item{metric}{The metric used for scaling covariates (default "E" for Euclidean).}

\item{catEncoding}{Optional list of categorical encoding metadata returned by
\code{encodeCategories_internal}, or \code{NULL} if no categorical covariates
were present.}
}
\value{
An object of class AddiVortes.
}
\description{
A constructor for the AddiVortes class.
}
