% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortesFit.R
\name{new_AddiVortes}
\alias{new_AddiVortes}
\title{Create an AddiVortes Object}
\usage{
new_AddiVortes(
  posteriorTess,
  posteriorDim,
  posteriorSigma,
  posteriorPred,
  xCentres,
  xRanges,
  yCentre,
  yRange,
  inSampleRmse,
  mcmcLog
)
}
\arguments{
\item{posteriorTess}{A list of the posterior samples of the tessellations.}

\item{posteriorDim}{A list of the posterior samples of the dimensions.}

\item{posteriorSigma}{A list of the posterior samples of the error variance.}

\item{posteriorPred}{A list of the posterior samples of the predictions.}

\item{xCentres}{The centres of the covariates.}

\item{xRanges}{The ranges of the covariates.}

\item{yCentre}{The centre of the output values.}

\item{yRange}{The range of the output values.}

\item{inSampleRmse}{The in-sample RMSE.}

\item{mcmcLog}{A data frame logging each proposed tessellation move during
MCMC. Contains columns \code{iteration} (MCMC iteration index),
\code{tessellation} (tessellation component index), \code{move_type}
(character: one of "AD", "RD", "AC", "RC", "Change", "Swap"),
\code{log_alpha} (log acceptance probability; \code{-Inf} for proposals
auto-rejected due to empty cells), \code{accepted} (logical), and
\code{variable} (integer covariate index added or removed for "AD"/"RD"
moves; \code{NA} for all other move types).}
}
\value{
An object of class AddiVortes.
}
\description{
A constructor for the AddiVortes class.
}
