% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortesFit.R
\name{plot.AddiVortesFit}
\alias{plot.AddiVortesFit}
\title{Plot Method for AddiVortesFit}
\usage{
\method{plot}{AddiVortesFit}(aModel, x_train, y_train, ...)
}
\arguments{
\item{aModel}{An object of class \code{AddiVortesFit}, typically the result of a
call to \code{AddiVortes()}.}

\item{x_train}{A matrix of the original training covariates.}

\item{y_train}{A numeric vector of the original training true outcomes.}

\item{...}{Additional graphical parameters to be passed to the \code{plot} function
(e.g., \code{pch}, \code{cex}).}
}
\value{
This function is called for its side effect of creating a plot and returns
\code{NULL} invisibly.
}
\description{
Generates diagnostic plots for a fitted \code{AddiVortesFit} object.
This function creates a scatter plot of the true versus predicted values
for the training set to help visualise model fit.
}
\details{
The function internally calls \code{predict.AddiVortesFit} on the provided
training data to get the mean predictions. It then plots these predictions
against the true outcome values (\code{y_train}). A dashed red line with an
intercept of 0 and a slope of 1 is added to the plot to represent a perfect
prediction, making it easy to assess the model's accuracy.
}
\examples{
\dontrun{
# Assuming 'fit' is a trained object of class AddiVortesFit,
# and 'x_train_data', 'y_train_data' are your training datasets.

plot(fit, x_train = x_train_data, y_train = y_train_data)
}
}
