% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictAddiVortes.R
\name{predictAddiVortes}
\alias{predictAddiVortes}
\title{predictAddiVortes}
\usage{
predictAddiVortes(addiVortesModelFit, xNew, yNew, quantiles = c(0.025, 0.975))
}
\arguments{
\item{addiVortesModelFit}{The 'addiVortesModel' list object returned by the
addiVortes function. This contains posterior samples
and scaling parameters.}

\item{xNew}{A matrix of the covariates for the new test set. The number of
columns must match the original training data.}

\item{yNew}{A vector of the true output values for the new test set. Must have
the same number of observations as rows in \code{xNew}.}

\item{quantiles}{A numeric vector of quantiles to compute for the predictions.}
}
\value{
A list that contains:
\itemize{
\item \code{outOfSampleRmse}: The Root Mean Squared Error of the predictions on the
new test data.
\item \code{meanYhatNewUnscaled}: The mean predictions for the new test data,
unscaled to the original scale of the output variable.
\item \code{quantileYhatNewUnscaled}: A matrix of quantiles of the predictions for the
new test data, unscaled to the original output variable scale.
}
}
\description{
Utilises a fitted addiVortes model to make predictions on new test data and
calculates the Root Mean Squared Error (RMSE). This function aims to
reproduce the Out-of-Sample RMSE calculation performed internally by the
addiVortes fitting function.
}
\details{
This function requires the helper functions \code{applyScalingInternal} and
\code{testSetPrediction} (used by the original \code{addiVortes} function) to be
available in the R environment.
}
