% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print.params}
\alias{print.params}
\alias{print.term_params}
\alias{print.geometry_params}
\alias{print.structure_params}
\alias{print.general_params}
\title{Print a parameter group}
\usage{
\method{print}{term_params}(x, ...)

\method{print}{geometry_params}(x, ...)

\method{print}{structure_params}(x, ...)

\method{print}{general_params}(x, ...)
}
\arguments{
\item{x}{A parameter group from \code{\link[=term_params]{term_params()}}, \code{\link[=geometry_params]{geometry_params()}},
\code{\link[=structure_params]{structure_params()}} or \code{\link[=general_params]{general_params()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a parameter group
}
\examples{
print(term_params(tessellations = 40))
}
