% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{print.thiessen_control}
\alias{print.thiessen_control}
\title{Print a control object}
\usage{
\method{print}{thiessen_control}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"thiessen_control"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a control object
}
\examples{
print(thiessen_control(tessellations = 50))
}
