% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.thiessen_outcome}
\alias{print.thiessen_outcome}
\title{Print an outcome family}
\usage{
\method{print}{thiessen_outcome}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"thiessen_outcome"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an outcome family
}
\examples{
print(gaussian_outcome())
}
