% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiessen-package.R
\docType{package}
\name{thiessen-package}
\alias{thiessen-package}
\title{thiessen: Bayesian Additive Voronoi Tessellations}
\description{
Bayesian regression on a sum of Voronoi tessellations (AddiVortes; Stone and Gosling, 2025, \doi{10.1080/10618600.2024.2414104}), a variant of BART (Chipman, George and McCulloch, 2010, \doi{10.1214/09-AOAS285}) in which a cell is a region of the covariate space rather than a box. Provides the Gaussian, binary probit and heteroscedastic models, with methods for prediction, posterior intervals and summaries. Further outcome families are experimental and reach a fit only in a source build made with the environment variable 'THIESSEN_EXPERIMENTAL' set to 1. The sampler is the 'thiessen' Rust crate, built and linked from vendored sources.
}
\references{
Chipman, H. A., George, E. I. and McCulloch, R. E. (2010). BART: Bayesian
additive regression trees. \emph{The Annals of Applied Statistics} 4(1),
266-298. \doi{10.1214/09-AOAS285}

Stone, A. and Gosling, J. P. (2025). AddiVortes: (Bayesian) additive
Voronoi tessellations. \emph{Journal of Computational and Graphical
Statistics} 34(3), 859-871. \doi{10.1080/10618600.2024.2414104}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://l-thomson.github.io/thiessen/r/}
  \item \url{https://github.com/l-thomson/thiessen}
  \item Report bugs at \url{https://github.com/l-thomson/thiessen/issues}
}

}
\author{
\strong{Maintainer}: Leo Thomson \email{leo@feasibly.co.uk}

Authors:
\itemize{
  \item Leo Thomson \email{leo@feasibly.co.uk}
}

Other contributors:
\itemize{
  \item The authors of the vendored Rust crates (listed in inst/AUTHORS) [copyright holder]
}

}
\keyword{internal}
