\docType{package}
\name{thiessen-package}
\alias{thiessen-package}
\alias{thiessen}
\title{thiessen: Bayesian Additive Voronoi Tessellations}
\description{
Bindings to the 'thiessen' Rust crate, an implementation of AddiVortes,
Bayesian regression on a sum of Voronoi tessellations (Stone and Gosling,
2025, \doi{10.1080/10618600.2024.2414104}). This version builds and links
the core crate from vendored sources; the modelling interface follows.
}
\author{
Leo Thomson \email{leo@feasibly.co.uk}
}
\seealso{
\itemize{
  \item \url{https://github.com/l-thomson/thiessen}
  \item Report bugs at \url{https://github.com/l-thomson/thiessen/issues}
}
}
\keyword{internal}
