% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddiVortesFit.R
\name{traceplots}
\alias{traceplots}
\title{Trace Plot Diagnostics}
\usage{
traceplots(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Further arguments passed to methods.}
}
\value{
This function is called for its side effect of creating plots and returns
\code{NULL} invisibly.
}
\description{
Creates trace plots for MCMC diagnostics of a fitted model.
}
